Also changed LONGs to floats. Define is in mathtrans_intern.h.
git-svn-id: https://svn.aros.org/svn/aros/trunk@3959 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/mathtrans/spcos.c
@@ -20,11 +20,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, SPCos,
+      AROS_LH1(float, SPCos,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, fnum1, D0),
+      AROS_LHA(float, fnum1, D0),
 
 /*  LOCATION */
