+++ workbench/libs/mathtrans/splog.c
@@ -20,11 +20,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, SPLog,
+      AROS_LH1(float, SPLog,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, fnum1, D0),
+      AROS_LHA(float, fnum1, D0),
 
 /*  LOCATION */
