+++ workbench/libs/mathtrans/spsinh.c
@@ -18,11 +18,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, SPSinh,
+      AROS_LH1(float, SPSinh,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, fnum1 , D0),
+      AROS_LHA(float, fnum1 , D0),
 
 /*  LOCATION */
